requirements.dev.txt
pyyaml==6.0

pynvim==0.4.3
pyinstaller==5.9.0
